// extensions/payment-customization/generated/input.ts
export type CartPaymentMethodsTransformRunInput = { cart: { lines: Array<{ id: string }>, cost: { totalAmount: { amount: string, currencyCode: CurrencyCode } } }, paymentMethods: Array<{ id: string, name: string }>, paymentCustomization: { config: { jsonValue: unknown } | null } };